AAAAA

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: alaorden <alaorden@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/10/07 19:56:54 by alaorden          #+#    #+#             */
-/*   Updated: 2025/10/17 17:40:11 by alaorden         ###   ########.fr       */
+/*   Updated: 2025/11/04 15:31:15 by alaorden         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *x)
+size_t	ft_strlen(const char *x)
 {
 	int	i;
